+++ proton-mail-uniffi/CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
     - Added `message_id_to_open` field into `WatchedConversation`
 
+### Removed
+
+  - Removed `load_remote_conversation()` as it is no longer usable with the
+    removal of remote IDs.
+
 ## [0.11.10] - 2024-08-20
 
 ### Added